refactor: Move create_dns_record to dns utils
